fix: made the requested changes, also working on router, not final
