+++ frontend/src/App.css
@@ -2,7 +2,3 @@ body {
     background-color: #1a1a1a;
     color: white;
 }
-
-header {
-    background-color: #444;
-}
