+++ frontend/src/App.tsx
@@ -1,17 +1,11 @@
-import { useState } from 'react'
 import './App.css'
-import Header from './Header'
-import Home from './Home'
-import Format1v1v from './Format-1v1v'
+import { Router } from './router/Router'
 
 function App() {
-  const [homePageState, setHomePageState] = useState(true)
 
   return (
     <>
-      <Header />
-       {/* <Home /> */}
-       <Format1v1v/>
+      <Router />
     </>
   )
 }
